agstack/pyproject.toml
[project]
name = "agstack"
version = "0.0.1"
description = ""
authors = [
    { name = "Name <Email>" }
]
license = { text = "MIT" }
requires-python = ">=3.10"

dependencies = [
    "agentstack[crewai]>=0.3.5",
    "embedchain>=0.1.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.23.2",
    "sqlalchemy",
    "psycopg2-binary",
    "agentops>=0.3.21",
    "pandas",
]
